# public/openapi.yaml
openapi: 3.1.0
info:
  title: 'Hexbatch core inspector'
  description: 'Testing api for hexbatch'
  termsOfService: 'https://hexbatch.com/core-inspector'
  contact:
    name: Hexbatch
    url: 'https://hexbatch.com'
  license:
    name: 'MIT License'
    identifier: MIT
  version: 0.1.0
servers:
  -
    url: 'http://localhost:8081'
    description: 'This tests on the local host'
paths:
...
  /api/v1/users/login:
    post:
      tags:
        - user
      summary: 'Logs the user in'
      description: 'This returns a token to use as basic auth in the other api calls'
      operationId: api.users.login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginParams'
      responses:
        '200':
          description: 'Login returns a token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: 'Bad credentials'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 422
                message: 'These credentials do not match our records'
  /api/v1/users/me:
    get:
      tags:
        - user
      summary: 'Returns logged-in user information'
      description: 'This returns user data about the person this authentication token is for'
      operationId: api.users.me
      responses:
        '200':
          description: 'Gives information about the logged in user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'When not logged in'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                message: Unauthenticated.
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    ErrorResponse:
      title: Error
      description: 'All errors will have this format'
      properties:
        status:
          title: 'The status of the error '
          description: 'This is normally a http code'
          type: integer
          example: 400
        message:
          title: 'The error message '
          description: 'Describes the problem'
          type: string
          example: 'You have not been assigned to this project'
      type: object
      example:
        status: 401
        message: 'Logged in user is not assigned to the project'
    LoginParams:
      title: 'Login Data'
      description: 'Login requires a username and password'
      required:
        - username
        - password
      properties:
        username:
          title: 'User name'
          type: string
          maxLength: 30
          minLength: 3
          example: dobby_1987
        password:
          title: Password
          type: string
          format: password
          example: beans_r_88good
      type: object
      example:
        username: will
        password: past-pass-power
        device_token: c2pIsrDb4v6cZS6u4uNYhx1v3sBibXci1_zI_BCb
    LoginResponse:
      title: 'Login response'
      description: 'Returns the token that is required by other api calls'
      properties:
        auth_token:
          title: 'Auth Token'
          type: string
          format: password
          example: 4|6u2HE63lbItRImAklzn96Axf1zUzXYYHOFotGLLR4fb81fa0
      type: object
    UserResponse:
      title: User
      description: 'This can describe a user or employee or someone who is both'
      properties:
        user_id:
          title: 'User Id'
          type:
            - integer
            - 'null'
          example: 58
        name:
          title: Name
          type:
            - string
            - 'null'
          example: Will
        username:
          title: Username
          type:
            - string
            - 'null'
          example: Will
        email:
          title: Email
          type:
            - string
            - 'null'
          format: email
          example: willtornweed@yarn.org
        user_tags:
          title: Tags
          description: $user_tags
          type: array
          items:
            type: string
        registration_date:
          title: 'Registration Date'
          type:
            - string
            - 'null'
          format: date-time
          example: '2015-08-12T00:00:00-05:00'
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'The user logs in to get a token, uses that as bearer authentication in the header'
      scheme: bearer
security:
  -
    bearerAuth: []
...